requirements.txt
lmdb==1.2.1
matplotlib==3.5.1
nltk==3.7
numpy==1.23.1
opencv_python==4.5.1.48
Pillow==9.2.0
six==1.16.0
timm==0.4.5
torch==1.10.0+cu113
torchvision==0.9.1
validators==0.18.2
